🪨 Integración completa de modelos 3D realistas para Sol y Asteroides
## ✨ Nuevos Modelos 3D Implementados

### 1. Sol Realista (SolRealista.java)
- ✅ Modelo 3D de alta calidad desde Sketchfab (Solrealista.obj)
- ✅ Textura realista del Sol (materialdelsol.png, 5MB)
- ✅ Shaders personalizados con efectos de plasma y lava
- ✅ Efectos de viento solar animados (SolarWinds.java)
- ✅ Escala proporcional: 3.5x respecto a la Tierra

### 2. Asteroides Realistas
- ✅ Modelo 3D optimizado (AsteroideRealista.obj - 600 triángulos)
- ✅ Textura rocosa gris realista (matasteroide.png)
- ✅ Reemplazo completo de:
  - Meteoritos que pasan por escena
  - Proyectiles disparados del arma
  - Meteoritos épicos (combo x10)
  - Meteoritos que impactan pantalla

## 🎨 Shaders y Efectos Visuales

### Shaders del Sol:
- `sol_vertex.glsl` / `sol_fragment.glsl` - Base del modelo
- `sol_effects_vertex.glsl` / `sol_effects_fragment.glsl` - Efectos de superficie
- `solar_winds_vertex.glsl` / `solar_winds_fragment.glsl` - Vientos solares
- `sol_lava_fragment.glsl` - Efectos de lava/plasma

### Shaders de Asteroides:
- `meteorito_simple_fragment.glsl` - Shader limpio para mostrar textura sin efectos
- Soporte para futuros efectos visuales preparado

## 📐 Ajustes de Escala y Tamaño

### Reducción de tamaños de asteroides (50-60%):
- **Meteoritos normales**: 0.04-0.28 → 0.02-0.14
- **Proyectiles de arma**: 0.08-0.32 → 0.03-0.12
- **Meteoritos pantalla**: 0.25-0.60 → 0.12-0.30
- **Meteoritos épicos**: 0.25-0.45 → 0.12-0.22
- Límite máximo: tamaño de la Luna (0.27 scale)

## 🔧 Mejoras Técnicas

### Sistema de Carga de Modelos:
- ✅ ObjLoader mejorado con soporte para materiales
- ✅ Carga de texturas separada del MTL
- ✅ Buffers estáticos compartidos para optimización de memoria
- ✅ Soporte para índices INT (modelos >32k vértices)
- ✅ Fan triangulation para caras con múltiples vértices

### Optimizaciones:
- Modelo 3D compartido entre todos los meteoritos (static buffers)
- Proyectiles con pooling usando mismo modelo 3D
- Texturas cargadas una sola vez por TextureManager

## 📚 Documentación Agregada

- `PARAMETROS_ASTEROIDES.md` - Guía completa de parámetros de asteroides
- `PARAMETROS_SOL_REALISTA.md` - Documentación del sistema del Sol
- `PARAMETROS_FORCEFIELD.md` - Parámetros del ForceField
- `SISTEMA_CARGA_MODELOS_BLENDER.md` - Guía de importación de modelos
- `GUIA_POSICIONES.md` - Referencia de posiciones en escena

## 🎯 Archivos Principales Modificados

### Java:
- `Meteorito.java` - Integración modelo 3D + shader simple para textura
- `MeteorShower.java` - Reducción de tamaños
- `Projectile.java` - Modelo 3D con pooling
- `SceneRenderer.java` - Integración SolRealista + TierraLiveHD
- `SolRealista.java` - NUEVO: Sistema completo del Sol 3D
- `SolarWinds.java` - NUEVO: Efectos de viento solar
- `TierraLiveHD.java` - NUEVO: Tierra HD con modelo 3D
- `ObjLoader.java` - Mejoras para modelos complejos

### Assets:
- `AsteroideRealista.obj/mtl` - Modelo + material asteroide
- `Solrealista.obj/mtl` - Modelo + material Sol
- `liveEarth.obj/mtl` - Modelo Tierra HD
- `matasteroide.png` - Textura asteroide (5.08 MB)
- `materialdelsol.png` - Textura Sol (5 MB)

## 🐛 Fixes Importantes

1. **Textura de asteroides**: Color blanco sin tinte para mostrar textura real
2. **Shader simple**: Elimina efectos de fuego procedurales, muestra textura pura
3. **Escalas proporcionales**: Sol 3.5x, Luna 0.27x referencia
4. **Blending correcto**: GL_SRC_ALPHA + depth mask para transparencias

## 🚀 Estado Actual

- ✅ Sol 3D realista funcionando con efectos
- ✅ Asteroides 3D con textura gris funcionando
- ✅ Tamaños balanceados y proporcionales
- ⏳ Preparado para agregar efectos de shader a asteroides

🎮 Ready para testing y ajustes finales de efectos visuales!

https://claude.com/claude-code

+++ app/src/main/java/com/secret/blackholeglow/UniverseBackground.java
@@ -97,18 +97,18 @@ public class UniverseBackground
         texCoordBuffer = mesh.uvBuffer;
 
         // Índices para triangulación
-        List<short[]> faces = mesh.faces;
+        List<int[]> faces = mesh.faces;  // ✅ int[] para compatibilidad con modelos grandes
         int tris = 0;
-        for (short[] f : faces) tris += f.length - 2;
+        for (int[] f : faces) tris += f.length - 2;
         indexCount = tris * 3;
         ShortBuffer ib = ByteBuffer
                 .allocateDirect(indexCount * Short.BYTES)
                 .order(ByteOrder.nativeOrder())
                 .asShortBuffer();
-        for (short[] f : faces) {
-            short v0 = f[0];
+        for (int[] f : faces) {
+            short v0 = (short) f[0];  // ✅ Cast a short (plano.obj es pequeño)
             for (int k = 1; k < f.length - 1; k++) {
-                ib.put(v0).put(f[k]).put(f[k + 1]);
+                ib.put(v0).put((short) f[k]).put((short) f[k + 1]);
             }
         }
         ib.position(0);
